+++ Sources/EudiWalletKit/Models/OpenId4VCIConfiguration.swift
@@ -1,3 +1,4 @@
+
 /*
 Copyright (c) 2026 European Commission
 
@@ -15,14 +16,14 @@ limitations under the License.
 */
 
 import Foundation
+import Copyable
 import CryptoKit
 import JOSESwift
-import SwiftyJSON
-import OpenID4VCI
 import MdocDataModel18013
 import MdocSecurity18013
+import OpenID4VCI
 import Security
-import Copyable
+import SwiftyJSON
 
 @Copyable
 public struct OpenId4VciConfiguration: Sendable {
@@ -77,7 +78,6 @@ public struct OpenId4VciConfiguration: Sendable {
 		self.dpopKeyOptions = dpopKeyOptions
 	}
 }
-
 extension CoseEcCurve {
 	var jwsAlgorithm: JWSAlgorithm? {
 		switch self {
@@ -88,7 +88,6 @@ extension CoseEcCurve {
 		}
 	}
 }
-
 extension OpenId4VciConfiguration {
 
 	static var supportedDPoPAlgorithms: Set<JWSAlgorithm> {
@@ -109,13 +108,16 @@ extension OpenId4VciConfiguration {
 			let ecCurve = keyOptions.curve
 			let ecAlgorithm = await secureArea.defaultSigningAlgorithm(ecCurve: keyOptions.curve)
 			guard let jwsAlg = ecCurve.jwsAlgorithm, algorithms.map(\.name).contains(jwsAlg.name) else {
-				throw PresentationSession.makeError(str: "Specified algorithm \(ecCurve.SECGName) not supported by server supported algorithms \(algorithms.map(\.name))") }
+				throw PresentationSession.makeError(str: "Specified algorithm \(ecCurve.SECGName) not supported by server supported algorithms \(algorithms.map(\.name))")
+			}
 			jwsAlgorithm = jwsAlg
 			let existingKeyInfo: KeyBatchInfo? = if let privateKeyId { try? await secureArea.getKeyBatchInfo(id: privateKeyId) } else { nil }
 			let hasCompatibleExistingKey = existingKeyInfo != nil && keyOptions.secureAreaName == existingKeyInfo?.secureAreaName && keyOptions.curve == ecCurve && existingKeyInfo?.usedCounts.count == 1
 			let existingPublicKey: CoseKey? = if hasCompatibleExistingKey, let privateKeyId { try? await secureArea.getPublicKey(id: privateKeyId, index: 0, curve: ecCurve) } else { nil }
-			let publicCoseKey: CoseKey = if let existingPublicKey { existingPublicKey } else {
-				(try await secureArea.createKeyBatch(id: keyId, credentialOptions: CredentialOptions(credentialPolicy: .rotateUse, batchSize: 1), keyOptions: keyOptions)).first! }
+			let publicCoseKey: CoseKey =
+				if let existingPublicKey { existingPublicKey } else {
+					(try await secureArea.createKeyBatch(id: keyId, credentialOptions: CredentialOptions(credentialPolicy: .rotateUse, batchSize: 1), keyOptions: keyOptions)).first!
+				}
 			let publicKeyJwk = try publicCoseKey.jwk
 			let unlockData = try await secureArea.unlockKey(id: keyId)
 			let signer = try SecureAreaSigner(secureArea: secureArea, id: keyId, index: 0, publicKey: publicKeyJwk.toJoseSwiftJWK(), curve: ecCurve, ecAlgorithm: ecAlgorithm, unlockData: unlockData)
@@ -131,8 +133,19 @@ extension OpenId4VciConfiguration {
 			jwsAlgorithm = JWSAlgorithm(name: algName)
 			logger.info("Signing algorithm for DPoP constructor to be used is: \(jwsAlgorithm.name)")
 			// EC supported bit sizes are 256, 384, or 521. RS256 is 2048 bits.
-			let bits: Int = switch jwsAlgorithm.name { case JWSAlgorithm(.ES256).name: 256; case JWSAlgorithm(.ES384).name: 384; case JWSAlgorithm(.ES512).name: 521; case JWSAlgorithm(.RS256).name: 2048; default: throw PresentationSession.makeError(str: "Unsupported DPoP algorithm: \(jwsAlgorithm.name)") }
-			let type: SecKey.KeyType = switch jwsAlgorithm.name { case JWSAlgorithm(.RS256).name: .rsa; default: .ellipticCurve }
+			let bits: Int =
+				switch jwsAlgorithm.name {
+				case JWSAlgorithm(.ES256).name: 256
+				case JWSAlgorithm(.ES384).name: 384
+				case JWSAlgorithm(.ES512).name: 521
+				case JWSAlgorithm(.RS256).name: 2048
+				default: throw PresentationSession.makeError(str: "Unsupported DPoP algorithm: \(jwsAlgorithm.name)")
+				}
+			let type: SecKey.KeyType =
+				switch jwsAlgorithm.name {
+				case JWSAlgorithm(.RS256).name: .rsa
+				default: .ellipticCurve
+				}
 			let privateKey: SecKey = if let privateKeyId, let pk = SecKey.getExistingKey(type: type, keyId: privateKeyId) { pk } else { try SecKey.createRandomKey(type: type, bits: bits, keyId: privateKeyId) }
 			signingKeyProxy = .secKey(privateKey)
 			publicKey = try KeyController.generateECDHPublicKey(from: privateKey)
@@ -144,15 +157,18 @@ extension OpenId4VciConfiguration {
 		}
 		return DPoPConstructor(algorithm: jwsAlgorithm, jwk: jwk, privateKey: signingKeyProxy)
 	}
+	
+	static let supportedCredentialReusePolicies: SupportedCredentialReusePolicies = .supported([.limitedTime, .onceOnly, .rotatingBatch])
 
 	func toOpenId4VCIConfig(credentialIssuerId: String, clientAttestationPopSigningAlgValuesSupported: [JWSAlgorithm]?) async throws -> OpenId4VCIConfig {
-		let client: Client = if let keyAttestationsConfig, clientAttestationPopSigningAlgValuesSupported != nil {
-			try await makeAttestationClient(config: keyAttestationsConfig, credentialIssuerId: credentialIssuerId, algorithms: clientAttestationPopSigningAlgValuesSupported)
-		} else {
-			.public(id: clientId)
-		}
-		let clientAttestationPoPBuilder: ClientAttestationPoPBuilder? = if keyAttestationsConfig != nil { DefaultClientAttestationPoPBuilder() } else { nil}
-		return OpenId4VCIConfig(client: client, authFlowRedirectionURI: authFlowRedirectionURI, authorizeIssuanceConfig: authorizeIssuanceConfig, requirePAR: requirePAR, clientAttestationPoPBuilder: clientAttestationPoPBuilder, issuerMetadataPolicy: issuerMetadataPolicy, requireDpop: requireDpop)
+		let client: Client =
+			if let keyAttestationsConfig, clientAttestationPopSigningAlgValuesSupported != nil {
+				try await makeAttestationClient(config: keyAttestationsConfig, credentialIssuerId: credentialIssuerId, algorithms: clientAttestationPopSigningAlgValuesSupported)
+			} else {
+				.public(id: clientId)
+			}
+		let clientAttestationPoPBuilder: ClientAttestationPoPBuilder? = if keyAttestationsConfig != nil { DefaultClientAttestationPoPBuilder() } else { nil }
+		return OpenId4VCIConfig(client: client, authFlowRedirectionURI: authFlowRedirectionURI, authorizeIssuanceConfig: authorizeIssuanceConfig, requirePAR: requirePAR, clientAttestationPoPBuilder: clientAttestationPoPBuilder, issuerMetadataPolicy: issuerMetadataPolicy, requireDpop: requireDpop, supportedCredentialReusePolicies: Self.supportedCredentialReusePolicies)
 	}
 
 	private func makeAttestationClient(config: KeyAttestationConfiguration, credentialIssuerId: String, algorithms: [JWSAlgorithm]?) async throws -> Client {
@@ -169,7 +185,7 @@ extension OpenId4VciConfiguration {
 		return client
 	}
 
-		/// Generates a deterministic key alias based on the CredentialIssuerId.
+	/// Generates a deterministic key alias based on the CredentialIssuerId.
 	///
 	/// This ensures the same key is reused for the same issuer across sessions. The alias is
 	/// generated by: